+++ src/main/java/by/vstu/isit/documentprocessor/repositories/DocpackageRepository.java
@@ -4,6 +4,6 @@ import by.vstu.isit.documentprocessor.entities.Docpackage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface DocpackageRepository extends JpaRepository<Docpackage, Integer>, JpaSpecificationExecutor<Docpackage> {
+public interface DocpackageRepository extends JpaRepository<Docpackage, Long>, JpaSpecificationExecutor<Docpackage> {
 
 }
