+++ src/main/java/by/vstu/isit/documentprocessor/repositories/FuncRepository.java
@@ -4,6 +4,6 @@ import by.vstu.isit.documentprocessor.entities.Func;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface FuncRepository extends JpaRepository<Func, Integer>, JpaSpecificationExecutor<Func> {
+public interface FuncRepository extends JpaRepository<Func, Long>, JpaSpecificationExecutor<Func> {
 
 }
