+++ src/main/java/by/vstu/isit/documentprocessor/repositories/OperRepository.java
@@ -4,6 +4,6 @@ import by.vstu.isit.documentprocessor.entities.Oper;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface OperRepository extends JpaRepository<Oper, Integer>, JpaSpecificationExecutor<Oper> {
+public interface OperRepository extends JpaRepository<Oper, Long>, JpaSpecificationExecutor<Oper> {
 
 }
