+++ src/main/java/by/vstu/isit/documentprocessor/repositories/TypeOperFuncRepository.java
@@ -4,6 +4,6 @@ import by.vstu.isit.documentprocessor.entities.TypeOperFunc;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface TypeOperFuncRepository extends JpaRepository<TypeOperFunc, Integer>, JpaSpecificationExecutor<TypeOperFunc> {
+public interface TypeOperFuncRepository extends JpaRepository<TypeOperFunc, Long>, JpaSpecificationExecutor<TypeOperFunc> {
 
 }
