+++ src/main/java/by/vstu/isit/documentprocessor/repositories/TypeOperRepository.java
@@ -4,6 +4,6 @@ import by.vstu.isit.documentprocessor.entities.TypeOper;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface TypeOperRepository extends JpaRepository<TypeOper, Integer>, JpaSpecificationExecutor<TypeOper> {
+public interface TypeOperRepository extends JpaRepository<TypeOper, Long>, JpaSpecificationExecutor<TypeOper> {
 
 }
